Refector: 프로필 코드 정리
