+++ src/components/ProfilePage/Profile/ProfileCover/ProfileCover.module.scss
@@ -1,0 +1,113 @@
+@use "@/styles/globals.scss" as *;
+@use "sass:color";
+
+.backgroundImage {
+  width: calc(100vw - 254px - 1px);
+  height: 400px;
+  position: absolute;
+  right: 0;
+  display: flex;
+  flex-direction: row;
+  align-items: flex-end;
+  justify-content: center;
+  overflow-x: hidden;
+  z-index: 2;
+
+  @include Size("tablet") {
+    width: calc(100vw - 80px - 1px);
+    height: 300px;
+  }
+
+  @include Size("mobile") {
+    height: 240px;
+    width: 100%;
+  }
+}
+
+.backgroundDefaultImageContainer {
+  background-color: $primary4;
+  width: 100%;
+  height: 200px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+}
+
+.backgroundDefaultImage {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.backgroundAddMessage {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  color: $gray0;
+  background-color: rgba(255, 255, 255, 0.1);
+  cursor: pointer;
+  z-index: 2;
+  padding: 16px 24px;
+  border-radius: 8px;
+}
+
+.coverBtns {
+  position: absolute;
+  right: 16px;
+  bottom: 16px;
+  z-index: 3;
+  display: flex;
+  align-items: center;
+  gap: 8px;
+
+  .coverEditBtn {
+    border: 1px solid $gray30;
+    color: $gray70;
+    background-color: $gray0;
+    padding: 11px;
+    border-radius: 50px;
+    @include Btn2;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    gap: 11px;
+    cursor: pointer;
+    width: 48px;
+    height: 48px;
+
+    @include Size("mobile") {
+      @include Btn3;
+      gap: 4px;
+      width: 40px;
+      height: 40px;
+    }
+
+    &:hover {
+      background-color: color.adjust($gray0, $lightness: 4%);
+    }
+
+    &:active {
+      background-color: color.adjust($gray0, $lightness: 8%);
+    }
+  }
+}
+
+.gradientOverlay {
+  position: absolute;
+  top: 60;
+  left: 258;
+  z-index: 1;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(180deg, rgba(255, 255, 255, 0) 40%, rgba(0, 0, 0, 0.4) 100%);
+}
+
+.hidden {
+  display: none;
+}
