+++ src/components/ProfilePage/ProfilePage.module.scss
@@ -7,23 +7,6 @@
   flex-direction: column;
   align-items: center;
 
-  .center {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: 100%;
-    max-width: $max-width-content;
-    padding: 0 32px;
-
-    @include Size("mobile") {
-      padding: 0;
-    }
-
-    @include Size("tablet") {
-      padding: 0;
-    }
-  }
-
   .feedContainer {
     width: 100%;
     display: flex;
